cDay5: Find length of set
Imagine you are managing a list of items purchased by different customers in an e-commerce application. Some customers may purchase the same items, resulting in duplicates in the list. You want to find out how many unique items were purchased in total. To do this, you can use a Python program to convert the list into a set (which automatically removes duplicates) and then find the length of the set, representing the number of unique items.